Create file for handling special key cases
