+++ src/lib/device/key_handlers.h
@@ -1,0 +1,8 @@
+#ifndef KEY_HANDLERS_H
+#define KEY_HANDLERS_H
+
+#include "keyboard.h"
+
+void specialHandler(KeyPress out);
+
+#endif
